fix(app-hooks/richdocuments): Workaround oxooldocuments breaks when the backend data isn't available.
The failover mechanism requires the backend capability/discovery data fetched before the app is launched.

This patch workarounds it by fetching the data using the in-container editor service's URL(the reverse proxy isn't working at the moment so it cannot be access via the web service) and place them to the proper path which will then be loaded by the app.

Fixes Mantis#2042.

Signed-off-by: Buo-ren Lin (OSSII) <buoren.lin@ossii.com.tw>